+++ Src/FluentAssertions.Json/JsonAssertionOptions.cs
@@ -4,7 +4,7 @@ using FluentAssertions.Equivalency;
 namespace FluentAssertions.Json
 {
     /// <summary>
-    /// Represents the run-time type-specific behavior of a json structural equivalency assertion. It is the equivalent of the class EquivalencyAssertionOptions in FluentAssertion
+    /// Represents the run-time type-specific behavior of a JSON structural equivalency assertion. It is the equivalent of <see cref="FluentAssertions.Equivalency.EquivalencyAssertionOptions{T}"/>
     /// </summary>
     public sealed class JsonAssertionOptions<T> : EquivalencyAssertionOptions<T> , IJsonAssertionOptions<T>
     {
