+++ Tests/FluentAssertions.Json.Specs/JTokenAssertionsSpecs.cs
@@ -1,6 +1,5 @@
 using System;
 using System.Collections.Generic;
-using FluentAssertions.Equivalency;
 using FluentAssertions.Formatting;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
@@ -454,13 +453,12 @@ namespace FluentAssertions.Json.Specs
             //-----------------------------------------------------------------------------------------------------------
             var actual = JToken.Parse("{ \"id\": 1.1232 }");
             var expected = JToken.Parse("{ \"id\": 1.1235 }");
-            const double precision = 1e-3;
             //-----------------------------------------------------------------------------------------------------------
             // Act & Assert
             //-----------------------------------------------------------------------------------------------------------
             actual.Should().BeEquivalentTo(expected,
                 options => options
-                .Using<double>(d => d.Subject.Should().BeApproximately(d.Expectation, precision))
+                .Using<double>(d => d.Subject.Should().BeApproximately(d.Expectation, 1e-3))
                 .WhenTypeIs<double>());           
         }
         [Fact]
@@ -471,34 +469,16 @@ namespace FluentAssertions.Json.Specs
             //-----------------------------------------------------------------------------------------------------------
             var actual = JToken.Parse("{ \"id\": 1.1232 }");
             var expected = JToken.Parse("{ \"id\": 1.1235 }");
-            const double precision = 1e-5;
             //-----------------------------------------------------------------------------------------------------------
             // Act & Assert
             //-----------------------------------------------------------------------------------------------------------
             
             actual.Should().
                 Invoking(x => x.BeEquivalentTo(expected,options => options
-                .Using<double>(d => d.Subject.Should().BeApproximately(d.Expectation, precision))
+                .Using<double>(d => d.Subject.Should().BeApproximately(d.Expectation, 1e-5))
                 .WhenTypeIs<double>()))
                 .Should().Throw<XunitException>()
                 .WithMessage("JSON document has a different value at $.id.*");
-        }
-        [Fact]
-        public void When_default_equivalencyassertionoptions_used_check_shoud_not_change_it()
-        {
-            using(new TempDefaultAssertionOptions(e => e
-                      .Using<double>(ctx => ctx.Subject.Should().BeApproximately(ctx.Expectation, 0.1))
-                      .WhenTypeIs<double>()))
-            {
-                // Arrange
-                var actual = JToken.Parse("{ \"id\": 1.1232 }");
-                var expected = JToken.Parse("{ \"id\": 1.1235 }");
-
-
-                // Act & Assert
-                actual.Should().BeEquivalentTo(expected, options => options);
-            }
-            
         }
         [Fact]
         public void When_the_value_of_a_property_contains_curly_braces_the_equivalency_check_should_not_choke_on_them()
@@ -1278,17 +1258,6 @@ namespace FluentAssertions.Json.Specs
             }, null);
             
             return output.ToString();
-        }
-        private sealed class TempDefaultAssertionOptions : IDisposable
-        {
-            public TempDefaultAssertionOptions(Func<EquivalencyAssertionOptions, EquivalencyAssertionOptions> config)
-            {
-                AssertionOptions.AssertEquivalencyUsing(config);
-            }
-            public void Dispose()
-            {
-                AssertionOptions.AssertEquivalencyUsing(_ => new EquivalencyAssertionOptions());
-            }
         }
     }
 }
